address pr comments from erikfrey
